requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
feedparser>=6.0.10
